+++ src/iw5/xsk/compiler.cpp
@@ -840,7 +840,7 @@ void compiler::emit_stmt_switch(const ast::stmt_switch::ptr& stmt, const block::
 
     emit_opcode(opcode::OP_endswitch, data);
 
-    auto offset =  7 * stmt->stmt->list.size();
+    auto offset = static_cast<std::uint32_t>(7 * stmt->stmt->list.size());
     function_->instructions.back()->size += offset;
     index_ += offset;
 
@@ -2660,7 +2660,7 @@ auto compiler::variable_stack_index(const ast::expr_identifier::ptr& name, const
         {
             if (blk->local_vars.at(i).init)
             {
-                return blk->local_vars_create_count - 1 - i;
+                return static_cast<std::uint8_t>(blk->local_vars_create_count - 1 - i);
             }
 
             throw comp_error(name->loc(), "local variable '" + name->value + "' not initialized.");
